Refactor savekey + add configs
The way autosave was constructing `SaveKey` wasnt very dynamic, and was too specific to be useful on pages with URLs that contain un-needed querystrings (params). This commit introduces more flexibility as well as an option to make autosave url (for SaveKey name) more flexible. In addition there are new configs available to allow an OP to adjust things such as input target identifier right from config.js (without requiring a plugin shim init for editor object). New settings available are:

 - `saveKeyPrefix: 'localstore_prefix',` <-- Optional `SaveKey` prefix, defaults to "autosave" for backwards compatibility
 - `saveKeyIgnoreParams: ['querystring_to_ignore', 'something_else', 'gclid'],` <-- Querystrings to scrub out of `SaveKey` URL
 - `saveKeyIgnoreProtocol: true,` <-- Whether or not to include the https:// in `SaveKey` name, defaults to false for backwards compatibility
 - `saveKeyDelimiter: ' | ',` <-- Optional `SaveKey` URL delimiter, defaults to "_" for backwards compatibility
 - `saveKeyAttribute: 'id'` <-- The unique attribute to target for `SaveKey` name, defaults to element "name" for backwards compatibility

The `SaveKey` is then constructed using those settings and pushed to config.